repair onboarding triggering
