+++ gui/src/App.tsx
@@ -35,7 +35,6 @@ import { SerialDetectionModal } from './components/SerialDetectionModal';
 import { VRCOSCSettings } from './components/settings/pages/VRCOSCSettings';
 import { TopBar } from './components/TopBar';
 import { TrackerSettingsPage } from './components/tracker/TrackerSettings';
-import { useConfig } from './hooks/config';
 import { OSCRouterSettings } from './components/settings/pages/OSCRouterSettings';
 import { useLocalization } from '@fluent/react';
 import * as os from '@tauri-apps/plugin-os';
@@ -52,6 +51,7 @@ import { useBreakpoint } from './hooks/breakpoint';
 import { VRModePage } from './components/vr-mode/VRModePage';
 import { InterfaceSettings } from './components/settings/pages/InterfaceSettings';
 import { error, log } from './utils/logging';
+import { AppLayout } from './AppLayout';
 
 export const GH_REPO = 'SlimeVR/SlimeVR-Server';
 export const VersionContext = createContext('');
@@ -59,92 +59,94 @@ export const DOCS_SITE = 'https://docs.slimevr.dev';
 export const SLIMEVR_DISCORD = 'https://discord.gg/slimevr';
 
 function Layout() {
-  const { loading } = useConfig();
-
-  if (loading) return <></>;
-
   const { isMobile } = useBreakpoint('mobile');
+
   return (
     <>
       <SerialDetectionModal></SerialDetectionModal>
       <VersionUpdateModal></VersionUpdateModal>
       <Routes>
-        <Route
-          path="/"
-          element={
-            <MainLayoutRoute isMobile={isMobile}>
-              <Home />
-            </MainLayoutRoute>
-          }
-        />
-        <Route
-          path="/vr-mode"
-          element={
-            <MainLayoutRoute isMobile={isMobile}>
-              <VRModePage />
-            </MainLayoutRoute>
-          }
-        />
-        <Route
-          path="/tracker/:trackernum/:deviceid"
-          element={
-            <MainLayoutRoute background={false} isMobile={isMobile}>
-              <TrackerSettingsPage />
-            </MainLayoutRoute>
-          }
-        />
-        <Route
-          path="/settings"
-          element={
-            <SettingsLayoutRoute>
-              <Outlet></Outlet>
-            </SettingsLayoutRoute>
-          }
-        >
-          <Route path="trackers" element={<GeneralSettings />} />
-          <Route path="serial" element={<Serial />} />
-          <Route path="osc/router" element={<OSCRouterSettings />} />
-          <Route path="osc/vrchat" element={<VRCOSCSettings />} />
-          <Route path="osc/vmc" element={<VMCSettings />} />
-          <Route path="interface" element={<InterfaceSettings />} />
+        <Route element={<AppLayout />}>
+          <Route
+            path="/"
+            element={
+              <MainLayoutRoute isMobile={isMobile}>
+                <Home />
+              </MainLayoutRoute>
+            }
+          />
+          <Route
+            path="/vr-mode"
+            element={
+              <MainLayoutRoute isMobile={isMobile}>
+                <VRModePage />
+              </MainLayoutRoute>
+            }
+          />
+          <Route
+            path="/tracker/:trackernum/:deviceid"
+            element={
+              <MainLayoutRoute background={false} isMobile={isMobile}>
+                <TrackerSettingsPage />
+              </MainLayoutRoute>
+            }
+          />
+          <Route
+            path="/settings"
+            element={
+              <SettingsLayoutRoute>
+                <Outlet />
+              </SettingsLayoutRoute>
+            }
+          >
+            <Route path="trackers" element={<GeneralSettings />} />
+            <Route path="serial" element={<Serial />} />
+            <Route path="osc/router" element={<OSCRouterSettings />} />
+            <Route path="osc/vrchat" element={<VRCOSCSettings />} />
+            <Route path="osc/vmc" element={<VMCSettings />} />
+            <Route path="interface" element={<InterfaceSettings />} />
+          </Route>
+          <Route
+            path="/onboarding"
+            element={
+              <OnboardingLayout>
+                <Outlet />
+              </OnboardingLayout>
+            }
+          >
+            <Route path="home" element={<HomePage />} />
+            <Route path="wifi-creds" element={<WifiCredsPage />} />
+            <Route path="connect-trackers" element={<ConnectTrackersPage />} />
+            <Route
+              path="calibration-tutorial"
+              element={<CalibrationTutorialPage />}
+            />
+            <Route
+              path="assign-tutorial"
+              element={<AssignmentTutorialPage />}
+            />
+            <Route path="trackers-assign" element={<TrackersAssignPage />} />
+            <Route path="enter-vr" element={<EnterVRPage />} />
+            <Route path="mounting/choose" element={<MountingChoose />}></Route>
+            <Route path="mounting/auto" element={<AutomaticMountingPage />} />
+            <Route path="mounting/manual" element={<ManualMountingPage />} />
+            <Route path="reset-tutorial" element={<ResetTutorialPage />} />
+            <Route
+              path="body-proportions/choose"
+              element={<ProportionsChoose />}
+            />
+            <Route
+              path="body-proportions/auto"
+              element={<AutomaticProportionsPage />}
+            />
+            <Route
+              path="body-proportions/manual"
+              element={<ManualProportionsPage />}
+            />
+            <Route path="done" element={<DonePage />} />
+          </Route>
+          <Route path="*" element={<TopBar></TopBar>}></Route>
         </Route>
-        <Route
-          path="/onboarding"
-          element={
-            <OnboardingLayout>
-              <Outlet></Outlet>
-            </OnboardingLayout>
-          }
-        >
-          <Route path="home" element={<HomePage />} />
-          <Route path="wifi-creds" element={<WifiCredsPage />} />
-          <Route path="connect-trackers" element={<ConnectTrackersPage />} />
-          <Route
-            path="calibration-tutorial"
-            element={<CalibrationTutorialPage />}
-          />
-          <Route path="assign-tutorial" element={<AssignmentTutorialPage />} />
-          <Route path="trackers-assign" element={<TrackersAssignPage />} />
-          <Route path="enter-vr" element={<EnterVRPage />} />
-          <Route path="mounting/choose" element={<MountingChoose />}></Route>
-          <Route path="mounting/auto" element={<AutomaticMountingPage />} />
-          <Route path="mounting/manual" element={<ManualMountingPage />} />
-          <Route path="reset-tutorial" element={<ResetTutorialPage />} />
-          <Route
-            path="body-proportions/choose"
-            element={<ProportionsChoose />}
-          />
-          <Route
-            path="body-proportions/auto"
-            element={<AutomaticProportionsPage />}
-          />
-          <Route
-            path="body-proportions/manual"
-            element={<ManualProportionsPage />}
-          />
-          <Route path="done" element={<DonePage />} />
-        </Route>
-        <Route path="*" element={<TopBar></TopBar>}></Route>
       </Routes>
     </>
   );
